Add files via upload
I obtained this data from Lehigh Universities website. The link to the page I used is: https://www2.lehigh.edu/student-life/clubs-and-organizations. On this page I found data regarding the amount of students actively participating in clubs & organizations at the university. My method for getting this data was going to Lehigh's website as I was sure they would have published this information. This data was valuable to me as someone who is involved with both club hockey and a member of a fraternity. Some of the key findings are that many students are active in the community and there are many options at the university when it comes to extracurricular activities/clubs. These numbers are probably very important to students who might be considering Lehigh as their future college and it could have large implications on their decision.